functions/requirements.txt
google-cloud-firestore>=2.11.1
google-cloud-storage>=2.10.0
firebase-admin>=6.1.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pandas>=2.0.0
